// backend/src/common/enums/restructure-type.enum.ts
export enum RestructureType {
  NORMAL_RESTRUCTURE = 'Normal Restructure',
  PRE_PAYMENT = 'Pre Payment',
  ADVANCE_PAYMENT = 'Advance Payment',
}

export enum RestructureStatus {
  INITIATED = 'Initiated',
  APPROVED = 'Approved',
  REJECTED = 'Rejected',
}

export enum InterestTreatment {
  CAPITALIZE = 'Capitalize',
  ADD_TO_FIRST_EMI = 'Add To First EMI',
  CARRY_FORWARD = 'Carry Forward',
}